A little better handling of the init file stuff

+++ redis.sh
@@ -12,8 +12,10 @@ fi
 rm /etc/redis/redis.conf
 mv files/redis.conf /etc/redis/redis.conf
 # upstart supervisor
-update-rc.d redis-server disable # disable init.d script for redis
-rm -f /etc/init.d/redis-server # Remove the existing init script
+if [ -f /etc/init.d/redis-server ]; then
+  update-rc.d redis-server disable # disable init.d script for redis
+  rm -f /etc/init.d/redis-server # Remove the existing init script
+fi
 
 # Create the upstart script for redis
 cat >/etc/init/redis-server.conf <<EOL
